move relativeTimeToDate to Utils
